+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 - [fix:fix nacos enabled condition bugs.](https://github.com/Tencent/spring-cloud-tencent/pull/795)
 - [feat:update getPort method of PolarisRegistration.](https://github.com/Tencent/spring-cloud-tencent/pull/805)
 - [fix:fix some compile output issues.](https://github.com/Tencent/spring-cloud-tencent/pull/812)
+- [fix:nacos & polaris discovery](https://github.com/Tencent/spring-cloud-tencent/pull/820)